+++ client/src/main/java/tachyon/hadoop/AbstractTFS.java
@@ -382,7 +382,7 @@ abstract class AbstractTFS extends FileSystem {
   public void initialize(URI uri, Configuration conf) throws IOException {
     super.initialize(uri, conf);
     LOG.info("initialize(" + uri + ", " + conf + "). Connecting to Tachyon: " + uri.toString());
-    UnderFileSystemUtils.addS3Credentials(conf);
+    Utils.addS3Credentials(conf);
     setConf(conf);
     mTachyonHeader = getScheme() + "://" + uri.getHost() + ":" + uri.getPort();
